Added react route links and dynamic dashboard for user or admin
