+++ src/App.js
@@ -1,8 +1,7 @@
-import logo from './logo.svg';
 import './App.css';
 import Login from './components/login';
-import axios from 'axios';
-import { useEffect, useState } from 'react';
+import Dashboard from './components/dashboard';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 const API_URL = 'http://localhost:3000/Login'
 
@@ -10,9 +9,18 @@ const API_URL = 'http://localhost:3000/Login'
 function App() {
 
   return (
-    <div className="App">
-      <Login/>
-    </div>
+    <Router>
+      <div className="App">
+        <Switch>
+          <Route exact path="/">
+            <Login/>
+          </Route>
+          <Route path="/dashboard">
+            <Dashboard/>
+          </Route>
+        </Switch>
+      </div>
+    </Router>
   );
 }
